+++ data/src/main/java/com/emotionstorage/data/remoteMediator/GetTimeCapsulesRemoteMediator.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.data.repoImpl
+package com.emotionstorage.data.remoteMediator
 
 import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadType
